+++ test/shared/src/main/scala/zio/test/mock/Expectation.scala
@@ -21,7 +21,7 @@ import scala.language.implicitConversions
 import zio.test.Assertion
 import zio.test.mock.Expectation.{ And, Chain, Or, Repeated }
 import zio.test.mock.ReturnExpectation.{ Fail, Succeed }
-import zio.test.mock.internal.{ MockException, MockRuntime, State }
+import zio.test.mock.internal.{ MockException, MockFactory, State }
 import zio.{ Has, IO, Layer, Managed, Tagged, ZLayer }
 
 /**
@@ -135,9 +135,9 @@ sealed trait Expectation[R <: Has[_]] { self =>
   private[test] val invocations: List[Int]
 
   /**
-   * Provided a `MockRuntime` constructs a layer with environment `R`.
+   * Provided a `Mock` constructs a layer with environment `R`.
    */
-  private[test] def mock: ZLayer[Has[MockRuntime], Nothing, R]
+  private[test] def mock: ZLayer[Has[Mock], Nothing, R]
 
   /**
    * Mock execution flag.
@@ -162,12 +162,12 @@ object Expectation {
     satisfied: Boolean,
     saturated: Boolean,
     invocations: List[Int]
-  )(val mock: ZLayer[Has[MockRuntime], Nothing, R])
+  )(val mock: ZLayer[Has[Mock], Nothing, R])
       extends Expectation[R]
 
   private[test] object And {
 
-    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[MockRuntime], Nothing, R]): And[R] =
+    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[Mock], Nothing, R]): And[R] =
       And(children, false, false, List.empty)(mock)
 
     private[test] object Items {
@@ -188,7 +188,7 @@ object Expectation {
     saturated: Boolean,
     invocations: List[Int]
   ) extends Expectation[R] {
-    def mock: ZLayer[Has[MockRuntime], Nothing, R] = method.mock
+    def mock: ZLayer[Has[Mock], Nothing, R] = method.mock
   }
 
   private[test] object Call {
@@ -209,12 +209,12 @@ object Expectation {
     satisfied: Boolean,
     saturated: Boolean,
     invocations: List[Int]
-  )(val mock: ZLayer[Has[MockRuntime], Nothing, R])
+  )(val mock: ZLayer[Has[Mock], Nothing, R])
       extends Expectation[R]
 
   private[test] object Chain {
 
-    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[MockRuntime], Nothing, R]): Chain[R] =
+    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[Mock], Nothing, R]): Chain[R] =
       Chain(children, false, false, List.empty)(mock)
 
     private[test] object Items {
@@ -232,12 +232,12 @@ object Expectation {
     satisfied: Boolean,
     saturated: Boolean,
     invocations: List[Int]
-  )(val mock: ZLayer[Has[MockRuntime], Nothing, R])
+  )(val mock: ZLayer[Has[Mock], Nothing, R])
       extends Expectation[R]
 
   private[test] object Or {
 
-    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[MockRuntime], Nothing, R]): Or[R] =
+    def apply[R <: Has[_]](children: List[Expectation[R]], mock: ZLayer[Has[Mock], Nothing, R]): Or[R] =
       Or(children, false, false, List.empty)(mock)
 
     private[test] object Items {
@@ -258,7 +258,7 @@ object Expectation {
     started: Int,
     completed: Int
   ) extends Expectation[R] {
-    def mock: ZLayer[Has[MockRuntime], Nothing, R] = child.mock
+    def mock: ZLayer[Has[Mock], Nothing, R] = child.mock
   }
 
   private[test] object Repeated {
@@ -315,7 +315,7 @@ object Expectation {
     ZLayer.fromManagedMany(
       for {
         state <- Managed.make(State.make(trunk))(State.checkUnmetExpectations)
-        env   <- (Mock.makeRuntime(state) >>> trunk.mock).build
+        env   <- (MockFactory.make(state) >>> trunk.mock).build
       } yield env
     )
 }
