+++ test/shared/src/main/scala/zio/test/mock/MockConsole.scala
@@ -19,7 +19,6 @@ package zio.test.mock
 import java.io.IOException
 
 import zio.console.Console
-import zio.test.mock.internal.MockRuntime
 import zio.{ Has, IO, UIO, ZLayer }
 
 object MockConsole {
@@ -32,7 +31,7 @@ object MockConsole {
   object PutStrLn extends Tag[String, Unit]
   object GetStrLn extends Tag[Unit, String]
 
-  private lazy val mock: ZLayer[Has[MockRuntime], Nothing, Console] =
+  private lazy val mock: ZLayer[Has[Mock], Nothing, Console] =
     ZLayer.fromService(mock =>
       new Console.Service {
         def putStr(line: String): UIO[Unit]   = mock(PutStr, line)
