+++ test/shared/src/main/scala/zio/test/mock/MockSystem.scala
@@ -17,7 +17,6 @@
 package zio.test.mock
 
 import zio.system.System
-import zio.test.mock.internal.MockRuntime
 import zio.{ Has, IO, UIO, ZLayer }
 
 object MockSystem {
@@ -30,7 +29,7 @@ object MockSystem {
   object Property      extends Tag[String, Option[String]]
   object LineSeparator extends Tag[Unit, String]
 
-  private lazy val mock: ZLayer[Has[MockRuntime], Nothing, System] =
+  private lazy val mock: ZLayer[Has[Mock], Nothing, System] =
     ZLayer.fromService(mock =>
       new System.Service {
         def env(variable: String): IO[SecurityException, Option[String]] = mock(Env, variable)
